requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/LLM
openai==1.3.7
anthropic==0.7.8

# Notion integration
notion-client==2.2.1

# Data processing
pandas==2.1.3
numpy==1.25.2

# Async support
aiohttp==3.9.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 